do_clear_area() vs map column 0
do_clear_area() runs a callback over each point in a square around
a center point.  When executed near the edge of the map, it clips
the square to avoid going over that edge.  But on the left side,
it was clipping to column 0 rather than 1, so running the callback
routine on column 0 even though that isn't part of the level.  This
bug doesn't seem to have caused any problems though.