cursesgraphics fix - S_engroom character
I normally force DECgraphics for tty and for curses so hadn't noticed
this before.  I recently let curses default to 'cursesgraphics' and the
engraving-in-room character was crossed horizontal and vertical bars,
the crosswall character.  It certainly stood out in the middle of a
room, but the emphasis seems out of proportion for "engraving here".

There is no 'epsilon' among the DEC line drawing characters.  Curses
might be able to render one, but not with DEC-style rendering of 0xEE.
Comment out S_engroom so that curses inherits the default backtick.