github issue #1274 - meta keys on MacOS+tty
I can't post comments to github but can commit something that will
result in an update of the issue.  Hopefully you'll see the commit.

Issue reported by RetepV:  using the Macbook option key as an alt
key and using that as a shift with a regular key does not generate
a meta key value for the regular key.

This behavior is not new.  I see the same thing on OSX 10.11.6.

The curses interface calls meta(,TRUE) and <option>+key produces
M-key as intended with it.  The tty interface can't use that, even
when nethack is linked against a curses library which contains it.
But it should be possible to reproduce the functionality, at least
for some of the paths through the confusing conditional mazes in
sys/share/unixtty.c and win/tty/termcap.c.

However, my attempts to do that have been unsuccessful.  Fetching
and sending the termcap 'mm' sequence ('smm' for terminfo) wasn't
enough, tinkering with the internal terminal settings wasn't enough,
nor combining the two.  I'm not a UNIX guru so the internals part
was guesswork.  And after trying to look at the ncurses source code
in bafflement I've thrown up my hands and given up.

But all is not lost.  NetHack's 'altmeta' option can be used to
simulate meta keys.  It's intended to support terminals that transmit
<esc> X when you type <alt>+X.  But it also allows manually typing
<esc> X (as two separate keys rather than as a shift) when nethack
is waiting for a command.  They will be combined into M-X.  It works
when positioning the cursor too, so that when using number_pad you
can use <esc> <digit> to move the cursor multiple spots at a time.

'altmeta' deliberately doesn't work when assigning names to objects
or making wishes or various other text entry situations, nor in menus
where <esc> works in the usual way.  A potential gotcha is that if
you type <esc> when nethack is waiting for a command, the game will
just sit there until you type another key.  That can be another <esc>
since <esc> <esc> is converted into plain <esc> rather than M-<esc>.