fix github issue #1278 - spellbook discovery
Issue reported by ars3niy:  having a pauper wizard #enhance bare-handed
fighting to basic caused the hero to learn all level 1 spellbooks.

Spellbook discovery for wizards when advancing skills is intentional,
to replace the old Luck-based chance of writing unknown books with
magic markers.  (The Luck bias for wizards still applies for scrolls.)

Discovering spellbooks when advancing non-spell skills does feel wrong.
Change it to only happen when advancing spell skills.

This isn't pauper-specific, it's just a lot more noticeable in that
state.

There may be better ways to cope with this, but for the time being I'm
marking the issue 'fixed'.

Fixes #1278